Updated to OpenGL 3.3 Core Profile (and probably broke everything in the process)

+++ src/net/ncplanner/plannerator/discord/play/smivilization/HutThingColorable.java
@@ -1,8 +1,8 @@
 package net.ncplanner.plannerator.discord.play.smivilization;
 import java.util.UUID;
-import net.ncplanner.plannerator.Renderer;
-import simplelibrary.config2.Config;
-import simplelibrary.image.Color;
+import net.ncplanner.plannerator.graphics.Renderer;
+import net.ncplanner.plannerator.config2.Config;
+import net.ncplanner.plannerator.graphics.image.Color;
 public abstract class HutThingColorable extends HutThing{
     private Color color;
     public HutThingColorable(UUID uuid, Hut hut, String name, String textureName, long price, Color defaultColor){
@@ -26,7 +26,7 @@ public abstract class HutThingColorable extends HutThing{
         color = new Color(config.get("rgb"));
     }
     @Override
-    public void draw(Renderer renderer, double left, double top, double right, double bottom){
+    public void draw(Renderer renderer, float left, float top, float right, float bottom){
         renderer.setColor(getColor());
         super.draw(renderer, left, top, right, bottom);
     }
