+++ src/net/ncplanner/plannerator/discord/play/smivilization/thing/Lamp.java
@@ -1,6 +1,6 @@
 package net.ncplanner.plannerator.discord.play.smivilization.thing;
 import java.util.UUID;
-import net.ncplanner.plannerator.Renderer;
+import net.ncplanner.plannerator.graphics.Renderer;
 import net.ncplanner.plannerator.discord.play.smivilization.Hut;
 import net.ncplanner.plannerator.discord.play.smivilization.HutThing;
 import net.ncplanner.plannerator.discord.play.smivilization.Wall;
@@ -14,7 +14,7 @@ public class Lamp extends HutThing{
         return new Lamp(uuid, hut);
     }
     @Override
-    public void draw(Renderer renderer, double left, double top, double right, double bottom){
+    public void draw(Renderer renderer, float left, float top, float right, float bottom){
         renderer.drawImage("/textures/smivilization/buildings/huts/gliese/furniture/lamp/lamp.png", left, top, right, bottom);
         if(on)renderer.drawImage("/textures/smivilization/buildings/huts/gliese/furniture/lamp/lamp glow.png", left, top, right, bottom);
     }
