+++ src/net/ncplanner/plannerator/multiblock/overhaul/fusion/Block.java
@@ -1,7 +1,7 @@
 package net.ncplanner.plannerator.multiblock.overhaul.fusion;
 import java.util.ArrayList;
 import java.util.function.Function;
-import net.ncplanner.plannerator.Renderer;
+import net.ncplanner.plannerator.graphics.Renderer;
 import net.ncplanner.plannerator.multiblock.Direction;
 import net.ncplanner.plannerator.multiblock.Multiblock;
 import net.ncplanner.plannerator.multiblock.configuration.AbstractPlacementRule;
@@ -13,8 +13,8 @@ import net.ncplanner.plannerator.planner.Core;
 import net.ncplanner.plannerator.planner.MathUtil;
 import net.ncplanner.plannerator.planner.StringUtil;
 import net.ncplanner.plannerator.planner.exception.MissingConfigurationEntryException;
-import simplelibrary.image.Color;
-import simplelibrary.image.Image;
+import net.ncplanner.plannerator.graphics.image.Color;
+import net.ncplanner.plannerator.graphics.image.Image;
 public class Block extends net.ncplanner.plannerator.multiblock.Block implements ITemplateAccess<net.ncplanner.plannerator.multiblock.configuration.overhaul.fusion.Block> {
     /**
      * MUST ONLY BE SET WHEN MERGING CONFIGURATIONS!!!
@@ -150,7 +150,7 @@ public class Block extends net.ncplanner.plannerator.multiblock.Block implements
         return tip;
     }
     @Override
-    public void renderOverlay(Renderer renderer, double x, double y, double width, double height, Multiblock multiblock){
+    public void renderOverlay(Renderer renderer, float x, float y, float width, float height, Multiblock multiblock){
         if(!isValid()){
             drawOutline(renderer, x, y, width, height, Core.theme.getBlockColorOutlineInvalid());
         }
@@ -170,7 +170,7 @@ public class Block extends net.ncplanner.plannerator.multiblock.Block implements
                 renderer.setColor(primaryColor, .125f);
                 renderer.fillRect(x, y, x+width, y+height);
                 renderer.setColor(primaryColor, .75f);
-                double border = width/8;
+                float border = width/8;
                 boolean top = cluster.contains(this.x, this.y, z-1);
                 boolean right = cluster.contains(this.x+1, this.y, z);
                 boolean bottom = cluster.contains(this.x, this.y, z+1);
@@ -209,7 +209,7 @@ public class Block extends net.ncplanner.plannerator.multiblock.Block implements
             }
             if(secondaryColor!=null){
                 renderer.setColor(secondaryColor, .75f);
-                double border = width/8;
+                float border = width/8;
                 boolean top = cluster.contains(this.x, this.y, z-1);
                 boolean right = cluster.contains(this.x+1, this.y, z);
                 boolean bottom = cluster.contains(this.x, this.y, z+1);
@@ -234,7 +234,7 @@ public class Block extends net.ncplanner.plannerator.multiblock.Block implements
         }
     }
     @Override
-    public void renderOverlay(Renderer renderer, double x, double y, double z, double width, double height, double depth, Multiblock multiblock, Function<Direction, Boolean> faceRenderFunc){
+    public void renderOverlay(Renderer renderer, float x, float y, float z, float width, float height, float depth, Multiblock multiblock, Function<Direction, Boolean> faceRenderFunc){
         if(!isValid()){
             drawOutline(renderer, x, y, z, width, height, depth, Core.theme.getBlockColorOutlineInvalid(), faceRenderFunc);
         }
@@ -243,7 +243,7 @@ public class Block extends net.ncplanner.plannerator.multiblock.Block implements
         }
         OverhaulFusionReactor.Cluster cluster = this.cluster;
         if(cluster!=null){
-            double border = width/16;
+            float border = width/16;
             Color primaryColor = null;
             if(cluster.netHeat>0){
                 primaryColor = Core.theme.getClusterOverheatingColor();
